third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/workers/xmlhttprequest-allowed-with-disabled-web-security-expected.txt

CONSOLE MESSAGE: line 11: PASS: Origin: <missing>
Test that disabling web security permits cross-origin XMLHttpRequest requests from a Worker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
